dataplane: extract cross-cutting counter accessors to maps_counters.go (#1686)
Same-package code motion only; method bodies byte-identical (verified).
Part of the #1686 split of pkg/dataplane/maps.go into domain accessor
files.
